Hi there, this is Jules. I've made a few changes to your code to resolve some issues that were causing problems with your automated checks.
Here’s a summary of what I did:

1.  **Resolved Redeclared Variables:** I refactored some code to centralize common elements, which fixed several "redeclared in this block" errors.

2.  **Fixed Undefined Type Errors:** I corrected some undefined type errors by updating import paths and adjusting how mock methods and types work with your libraries.

3.  **Updated MockUser Interface Compliance:** I added a method to ensure a mock component correctly implements its intended interface.

4.  **Corrected i18n Test Failures and Logic:** I fixed a bug in a localization function and updated the corresponding tests to match the corrected behavior.

5.  **Hardened `go-check.yml` Workflow:** I made adjustments to your continuous integration workflow to ensure that errors in various steps will correctly cause the workflow to fail, providing more accurate feedback. I kept a setting that allows all checks to complete and report their status independently.

These changes should make your automated checks more reliable and accurately reflect the state of your codebase.
