+++ internal/bot/handler/set.go
@@ -88,20 +88,8 @@ func (s *Set) Middlewares() []tb.MiddlewareFunc {
 	return nil
 }
 
-const (
-	SetFeedItemButtonUnique = "set_feed_item_btn"
-	feedSettingTmpl         = `
-{{ .L "set_tmpl_header_settings" }}
-{{ .L "set_tmpl_label_id" }} {{ .source.ID }}
-{{ .L "set_tmpl_label_title" }} {{ .source.Title }}
-{{ .L "set_tmpl_label_link" }} {{ .source.Link }}
-{{ .L "set_tmpl_label_updates" }} {{if ge .source.ErrorCount .Count }}{{ .L "set_tmpl_status_paused" }}{{else}}{{ .L "set_tmpl_status_active" }}{{end}}
-{{ .L "set_tmpl_label_interval" }} {{ .sub.Interval }} {{ .L "set_tmpl_unit_minutes" }}
-{{ .L "set_tmpl_label_notifications" }} {{if eq .sub.EnableNotification 0}}{{ .L "set_tmpl_status_off" }}{{else}}{{ .L "set_tmpl_status_on" }}{{end}}
-{{ .L "set_tmpl_label_telegraph" }} {{if eq .sub.EnableTelegraph 0}}{{ .L "set_tmpl_status_off" }}{{else}}{{ .L "set_tmpl_status_on" }}{{end}}
-{{ .L "set_tmpl_label_tags" }} {{if .sub.Tag}}{{ .sub.Tag }}{{else}}{{ .L "set_tmpl_status_none" }}{{end}}
-`
-)
+// SetFeedItemButtonUnique is defined in common.go
+// feedSettingTmpl is defined in common.go
 
 type SetFeedItemButton struct {
 	bot  *tb.Bot
@@ -150,13 +138,9 @@ func (r *SetFeedItemButton) Handle(ctx tb.Context) error {
 		return ctx.Edit(i18n.Localize(langCode, "set_err_user_not_subscribed"))
 	}
 
-	funcMap := template.FuncMap{
-		"L": func(key string, args ...interface{}) string {
-			return i18n.Localize(langCode, key, args...)
-		},
-	}
-	t := template.New("setting template").Funcs(funcMap)
-	_, err = t.Parse(feedSettingTmpl)
+	// Use common getTemplateFuncMap and feedSettingTmpl
+	t := template.New("setting template").Funcs(getTemplateFuncMap(langCode))
+	_, err = t.Parse(feedSettingTmpl) // feedSettingTmpl is now from common.go
 	if err != nil {
 		// Log error, return generic message
 		return ctx.Edit(i18n.Localize(langCode, "set_err_button_settings_error"))
@@ -172,67 +156,12 @@ func (r *SetFeedItemButton) Handle(ctx tb.Context) error {
 	return ctx.Edit(
 		text.String(),
 		&tb.SendOptions{ParseMode: tb.ModeHTML},
+		// Use genFeedSetBtn from common.go
 		&tb.ReplyMarkup{InlineKeyboard: genFeedSetBtn(ctx.Callback(), sub, source, langCode)},
 	)
 }
 
-func genFeedSetBtn(
-	c *tb.Callback, sub *model.Subscribe, source *model.Source, langCode string,
-) [][]tb.InlineButton {
-	setSubTagKey := tb.InlineButton{
-		Unique: SetSubscriptionTagButtonUnique,
-		Text:   i18n.Localize(langCode, "set_btn_tag_settings"),
-		Data:   c.Data,
-	}
-
-	var notificationTextKey string
-	if sub.EnableNotification == 1 {
-		notificationTextKey = "set_btn_disable_notifications"
-	} else {
-		notificationTextKey = "set_btn_enable_notifications"
-	}
-	toggleNoticeKey := tb.InlineButton{
-		Unique: NotificationSwitchButtonUnique,
-		Text:   i18n.Localize(langCode, notificationTextKey),
-		Data:   c.Data,
-	}
-
-	var telegraphTextKey string
-	if sub.EnableTelegraph == 1 {
-		telegraphTextKey = "set_btn_disable_telegraph"
-	} else {
-		telegraphTextKey = "set_btn_enable_telegraph"
-	}
-	toggleTelegraphKey := tb.InlineButton{
-		Unique: TelegraphSwitchButtonUnique,
-		Text:   i18n.Localize(langCode, telegraphTextKey),
-		Data:   c.Data,
-	}
-
-	var updatesTextKey string
-	if source.ErrorCount >= config.ErrorThreshold {
-		updatesTextKey = "set_btn_resume_updates"
-	} else {
-		updatesTextKey = "set_btn_pause_updates"
-	}
-	toggleEnabledKey := tb.InlineButton{
-		Unique: SubscriptionSwitchButtonUnique,
-		Text:   i18n.Localize(langCode, updatesTextKey),
-		Data:   c.Data,
-	}
-
-	feedSettingKeys := [][]tb.InlineButton{
-		{ // Row 1
-			toggleEnabledKey,
-			toggleNoticeKey,
-		},
-		{ // Row 2
-			toggleTelegraphKey,
-			setSubTagKey,
-		},
-	}
-	return feedSettingKeys
-}
+// genFeedSetBtn is defined in common.go
 
 func (r *SetFeedItemButton) Middlewares() []tb.MiddlewareFunc {
 	return nil
