+++ internal/bot/handler/set_subscription_tag_button.go
@@ -10,10 +10,8 @@ import (
 	"github.com/zintus/flowerss-bot/internal/i18n"
 )
 
-const (
-	SetSubscriptionTagButtonUnique = "set_set_sub_tag_btn"
-	// Use util.DefaultLanguage instead of local declaration
-)
+// SetSubscriptionTagButtonUnique is defined in common.go
+// Use util.DefaultLanguage instead of local declaration
 
 type SetSubscriptionTagButton struct {
 	bot *tb.Bot
